Add return_value_as<T>() for portable fixed-width integer returns
Fixed-width types (uint64_t, size_t, ptrdiff_t, etc.) are typedefs for
platform-specific fundamental types: unsigned long on LP64, unsigned long
long on LLP64. Callers couldn't portably read back a return value without
knowing which storage slot was used.

return_value_as<T>() and return_value_as_or_default<T>() dispatch through
the widest compatible getter (return_unsigned_long_long_int_value() for
unsigned types, return_long_long_int_value() for signed), which already
accept all narrower storage widths, so the correct value is retrieved
regardless of the underlying fundamental type the platform chose.

+++ tests/integration/MockSupport.c
@@ -28,13 +28,9 @@ void mutiny_all_mock_support_c_calls(void)
       ->with_bool_parameters("bool", 1)
       ->with_unsigned_int_parameters("unsigned", 1)
       ->with_long_int_parameters("long int", (long int)-1)
-      ->with_unsigned_long_int_parameters(
-          "unsigned long int", (unsigned long int)1
-      )
+      ->with_unsigned_long_int_parameters("unsigned long int", 1UL)
       ->with_long_long_int_parameters("long long int", (long long int)-1)
-      ->with_unsigned_long_long_int_parameters(
-          "unsigned long long int", (unsigned long long int)1
-      )
+      ->with_unsigned_long_long_int_parameters("unsigned long long int", 1ULL)
       ->with_double_parameters("double", 1.0)
       ->with_double_parameters_and_tolerance("doubleWithTolerance", 1.0, 1.0)
       ->with_string_parameters("string", "string")
@@ -48,15 +44,11 @@ void mutiny_all_mock_support_c_calls(void)
       ->actual_call("boo")
       ->with_int_parameters("integer", 1)
       ->with_bool_parameters("bool", 1)
-      ->with_unsigned_int_parameters("unsigned", 1)
-      ->with_long_int_parameters("long int", (long int)-1)
-      ->with_unsigned_long_int_parameters(
-          "unsigned long int", (unsigned long int)1
-      )
-      ->with_long_long_int_parameters("long long int", (long long int)-1)
-      ->with_unsigned_long_long_int_parameters(
-          "unsigned long long int", (unsigned long long int)1
-      )
+      ->with_unsigned_int_parameters("unsigned", 1U)
+      ->with_long_int_parameters("long int", -1L)
+      ->with_unsigned_long_int_parameters("unsigned long int", 1UL)
+      ->with_long_long_int_parameters("long long int", -1LL)
+      ->with_unsigned_long_long_int_parameters("unsigned long long int", 1ULL)
       ->with_double_parameters("double", 1.0)
       ->with_double_parameters("doubleWithTolerance", 0.0)
       ->with_string_parameters("string", "string")
